### 2025-08-12 #8784 @editorialbot
[PRE REVIEW]: pychoco: all-inclusive Python bindings for the Choco-solver constraint programming library
**Submitting author:** @dimitri-justeau (<a href="http://orcid.org/0000-0003-4129-0764">Dimitri Justeau-Allaire</a>)
**Repository:** https://github.com/chocoteam/pychoco
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.2.3

Thanks for submitting your paper to JOSS @dimitri.justeau. **Currently, there isn't a JOSS editor assigned** to your paper.

@dimitri.justeau if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
