### 2025-08-12 #8784 @danielskatz
> I am not sure how to proceed with this, as the license is not OSI-approved, and the submitter did not provide a GitHub username, nor does a search of GitHub find the submitter.

I found the submitter and fixed the typo in their username.
