### 2025-08-12 #8784 @dimitri-justeau
Hi @danielskatz , sorry if there was a problem with my GitHub username. About the license, it is (by default) the same as Choco-solver. I would be in favor of switching to GPL-3, but Charles Prud'homme (@cprudhom , https://github.com/cprudhom) needs to be included in the loop. Can we include him ? Thank you.
