### 2025-08-12 #8784 @cprudhom
Hi
I think we should change the license to BSD-3 clauses, it is close to the previous one but OSI approved.
