### 2025-08-18 b967857
Update bib
