+++ paper/paper.bib
@@ -1,40 +1,38 @@
 @book{rossi_2006,
-  editor    = {Francesca Rossi and
-               Peter van Beek and
-               Toby Walsh},
-  title     = {Handbook of Constraint Programming},
-  series    = {Foundations of Artificial Intelligence},
-  volume    = {2},
-  publisher = {Elsevier},
-  year      = {2006},
-  url       = {https://www.sciencedirect.com/science/bookseries/15746526/2},
-  isbn      = {978-0-444-52726-4},
+  editor={Francesca Rossi and Peter van Beek and Toby Walsh},
+  title={Handbook of Constraint Programming},
+  series={Foundations of Artificial Intelligence},
+  volume={2},
+  publisher={Elsevier},
+  year={2006},
+  url={https://www.sciencedirect.com/science/bookseries/15746526/2},
+  isbn={978-0-444-52726-4},
 }
 @incollection{nethercote_2007,
-  title = {{{MiniZinc}}: {{Towards}} a {{Standard CP Modelling Language}}},
-  shorttitle = {{{MiniZinc}}},
-  booktitle = {Principles and {{Practice}} of {{Constraint Programming}} -- {{CP}} 2007},
-  author = {Nethercote, Nicholas and Stuckey, Peter J. and Becket, Ralph and Brand, Sebastian and Duck, Gregory J. and Tack, Guido},
-  editor = {Bessi{\`e}re, Christian},
-  year = {2007},
-  volume = {4741},
-  pages = {529--543},
-  publisher = {Springer Berlin Heidelberg},
-  address = {Berlin, Heidelberg},
-  doi = {10.1007/978-3-540-74970-7_38},
-  isbn = {978-3-540-74969-1},
+  title={{{MiniZinc}}: {{Towards}} a {{Standard CP Modelling Language}}},
+  shorttitle={{{MiniZinc}}},
+  booktitle={Principles and {{Practice}} of {{Constraint Programming}} -- {{CP}} 2007},
+  author={Nethercote, Nicholas and Stuckey, Peter J. and Becket, Ralph and Brand, Sebastian and Duck, Gregory J. and Tack, Guido},
+  editor={Bessi{\`e}re, Christian},
+  year={2007},
+  volume={4741},
+  pages={529--543},
+  publisher={Springer Berlin Heidelberg},
+  address={Berlin, Heidelberg},
+  doi={10.1007/978-3-540-74970-7_38},
+  isbn={978-3-540-74969-1},
 }
 @article{audemard_2020,
-  title = {{{XCSP3}} and Its Ecosystem},
-  author = {Audemard, Gilles and Boussemart, Fr{\'e}d{\'e}ric and Lecoutre, Christophe and Piette, C{\'e}dric and Roussel, Olivier},
-  year = {2020},
-  month = apr,
-  journal = {Constraints},
-  volume = {25},
-  number = {1},
-  pages = {47--69},
-  issn = {1572-9354},
-  doi = {10.1007/s10601-019-09307-9},
+  title={{{XCSP3}} and Its Ecosystem},
+  author={Audemard, Gilles and Boussemart, Fr{\'e}d{\'e}ric and Lecoutre, Christophe and Piette, C{\'e}dric and Roussel, Olivier},
+  year={2020},
+  month=apr,
+  journal={Constraints},
+  volume={25},
+  number={1},
+  pages={47--69},
+  issn={1572-9354},
+  doi={10.1007/s10601-019-09307-9},
 }
 @inproceedings{guns_2019,
     title={Increasing modeling language convenience with a universal n-dimensional array, CPpy as python-embedded example},
@@ -44,87 +42,90 @@
     year={2019},
 }
 @article{prudhomme_2022,
[…]
