### 2025-08-19 #8784 @IgnaceBleukx
Hi! Yes I would be happy to review this submission for JOSS.
