### 2025-08-21 #8847 @editorialbot
[REVIEW]: pychoco: all-inclusive Python bindings for the Choco-solver constraint programming library
**Submitting author:** @dimitri.justeau (<a href="http://orcid.org/0000-0003-4129-0764">Dimitri Justeau-Allaire</a>)
**Repository:** https://github.com/chocoteam/pychoco
**Branch with paper.md** (empty if default branch): master
**Version:** pychoco-0.2.4
