### 2025-08-26 b808b2e
Update paper and notebooks
