### 2025-08-26 #8847 @dimitri-justeau
@IgnaceBleukx & @skadio , I updated the paper according to skadio's comments in the pre-review. I also added a figure depicting the pychoco Cheat Sheet, with a link to it. Sorry for the 3 generations, I had small problems to fix.
