### 2025-09-16 edee05a
Improvments thant to Joss review skadio
