### 2025-09-16 #8847 @skadio
@dimitri-justeau  thank you for the updates and improvements over the previous version. Adding the Python examples next to Java is a great way to go, and I hope you continue to build that out over time --so we have everything in one place; tutorials and examples.

The Readme details are very nice too. I understand the complication on custom goals, and in the meantime, documenting the generic capabilities should help the reader.

This is an area that I've encountered several times where most ML/Python practitioners end up writing their own heuristics instead of relying on a constraint solver like PyChoco with a declarative search goal -- simply because, they cannot get the solution they want out of the system. Here is a huge +1 on your efforts enabling this in the future.

Thanks again for the updates --the final version looks good to me!
