### 2025-09-24 #8847 @IgnaceBleukx
Thank you for you contribution @dimitri-justeau
I much appreciate the effort of making the python interface to the Choco solver, allowing users access to a "true" CP solver with many global constraints implemented. From a modelers perspective, this is one of the main contributions the pychoco package brings to the Python world.

Below I have some comments, though none of them are blocking for publication.

#### 1. Alternative solvers require separate installation of the solver
I feel this statement is somewhat outdated currently. While indeed most commercial solvers require a user to install a separate binary that needs to be linked to the Python interface, most open-source solvers are pip-installable to the best of my knowledge. E.g., in CPMpy, we have a requirement of open-source solvers being pip-installable (OR-Tools, Glasgow Constraint Solver, Pumpkin...). Though indeed, Choco is to my knowledge the only java-based solver which is this easy to install from a user-perspective.

#### 2. Documentation of global constraints
As far as I can tell, there is no exhaustive list available listing **all** constraints supported by the solver? Hence users need to go through the source code (or rely on the auto-complete of the IDE) to find which constraints are available.

#### 3. Link to notebook in paper
Currently, the link to the notebook listing examples is broken in the pdf. It links to the non-existing url: https://github.com/chocoteam/pychoco/tree/master/docs/notebooks

The remainder of my questions are mostly technical and targeted to the pychoco software itself, rather than the paper

#### 4. Numpy support?
I very much appreciate the typing hints being implemented in the pychoco interface. As it currently stands, it seems that numpy-constants are not allowed to be used in constraints? I think that for using pychoco in combination with a data science pipeline, this would be great to have from a user-perspective. Of course the user can simply cast numpy int/bools to Python-native objects but it would be great if the interface does this for them.

#### 5. Choco solver version
On the readme of pychoco, you mention the current solver version shipped with pychoco is version 4.10.17, though the "upstream" java version is currently at 5.0.0. Is there a special reason why these are not synced? It would probably be beneficial to release a new pychoco version together with every Choco-solver version?

Thanks again for the contribution! I think it is a great step in making CP-technology more available to non-experts.
