### 2025-09-25 0c7c681
Update to Choco-solver 4.10.18
