### 2025-09-25 a337f34
Update to Choco-solver 4.10.18
