### 2025-09-26 #8847 @IgnaceBleukx
@danielskatz yes! I am positive this paper is ready to be accepted and published!
