### 2025-09-27 b821a28
Update paper.md
