### 2025-09-28 #8847 @danielskatz
@dimitri-justeau - Can you change the metadata of the archive to indicate the version tag?
