### 2025-09-28 #8847 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
