### 2025-09-29 #8847 @dimitri-justeau
Everything looks good, thank you 👌

> [@editorialbot](https://github.com/editorialbot) recommend-accept
>
> [@dimitri-justeau](https://github.com/dimitri-justeau) - also, please check the proof that this will generate and confirm it looks good to you
