### 2025-09-29 #8847 @dimitri-justeau
@danielskatz , sorry to identify this this late, @cprudhom detected a typo in the end of the "Design" section:

"Finally, in addition to **an** comprehensive code documentation"

I just corrected it in both the main and paper branches, and also updated the figure because the "Cheat Sheet" text was in white and therefore invisible. I also updated the downloads count.

Is that ok ?
