### 2025-09-29 1f2f9dc
Fix typo in paper
