### 2025-09-29 #8847 @danielskatz
why don't we just drop the paper branch at this point. I'll just build from main...
