### 2025-09-29 #8847 @danielskatz
Congratulations to @dimitri-justeau and @cprudhom on your JOSS publication!!

And thanks to @IgnaceBleukx and @skadio for reviewing!
JOSS is completely dependent on volunteers, and we couldn't do this without you

(note: for me, the DOI resolves, but https://www.theoj.org/joss-papers/joss.08847/10.21105.joss.08847.pdf isn't loading, so I'll keep this open until everything works)
