### 2025-09-30 #8847 @dimitri-justeau
We are very happy to have this paper accepted on JOSS, thank you @danielskatz , @skadio , @IgnaceBleukx , and of course the devoted @editorialbot for making this fluent and pleasing :-)
